CONSOLE-3769: Refactor weird `getResourceSidebarSamples`
