+++ frontend/packages/console-shared/src/utils/index.ts
@@ -15,7 +15,6 @@ export * from './alert-utils';
 export * from './operator-utils';
 export * from './helm-utils';
 export * from './hpa-utils';
-export * from './sample-utils';
 export * from './multiselectdropdown';
 export * from './annotations';
 export * from './yup-validations';
