+++ frontend/public/components/sidebars/resource-sidebar.tsx
@@ -12,7 +12,7 @@ import {
 } from './resource-sidebar-samples';
 import { ExploreType } from './explore-type-sidebar';
 import { SimpleTabNav, Tab } from '../utils/simple-tab-nav';
-import { Sample } from '@console/shared/src/utils/sample-utils';
+import { Sample } from '@console/shared/src/hooks/useResourceSidebarSamples';
 import { Flex, FlexItem, Title } from '@patternfly/react-core';
 import PaneBody from '@console/shared/src/components/layout/PaneBody';
